+++ cypress/utils/index.js
@@ -1,1 +1,1 @@
-export { testInputValue } from './test_input_value.js';
+export { testInputValue } from './testInputValue.js';
