+++ src/js/util/getRandomNumber.js
